+++ src/main/java/com/daveestar/bettervanilla/enums/Permissions.java
@@ -21,6 +21,7 @@ public enum Permissions {
   TOGGLELOCATION("bettervanilla.togglelocation", Category.PLAYER),
   TOGGLECOMPASS("bettervanilla.togglecompass", Category.PLAYER),
   CHESTSORT("bettervanilla.chestsort", Category.PLAYER),
+  DOUBLE_DOOR("bettervanilla.doubledoor", Category.PLAYER),
   VEINMINER("bettervanilla.veinminer", Category.PLAYER),
   VEINCHOPPER("bettervanilla.veinchopper", Category.PLAYER);